+++ Search/EventListener/DeindexListener.php
@@ -31,7 +31,7 @@ class DeindexListener
 
     /**
      * Deindex subject from event.
-     * 
+     *
      * @param DeindexEvent $event
      */
     public function onDeindex(DeindexEvent $event)