+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 CHANGELOG
 =========
 
+* dev-develop
+    * FEATURE #80 Added composer handler for creating zend lucene directory
+    * FEATURE #74 Added support for resuming interupted reindexing tasks
+    * FEATURE #74 Deprecated `massive:search:index:rebuild` command in favor
+                  of `massive:search:reindex`
+    * FEATURE #74 [BC BREAK] Removed REINDEX event. Reindex providers must be
+      used instead.
+    * BUGFIX  #73 Removed typehint to avoid failing testcases in php7
+
 * 0.12.0 (2015-12-11)
     * HOTFIX  #69 Improve memory usage of rebuilding search indexes
     * HOTFIX  #72 Fixed null date-value